Refactor product service imports to use new directory structure
 • Move product-related service functions into separate directories
 • Update import paths in various files to reflect new locations
 • Remove product.ts and product.spec.ts as their contents are now in individual files
 • Add new tests for destroy, get, list-ids, list, remove-location-from-all, and upsert functions within the product service
 • Adjust user service product-related imports to match new product service structure
