+++ src/functions/customer/controllers/product/destroy.spec.ts
@@ -9,7 +9,7 @@ import {
 import { omitObjectIdProps } from "~/library/jest/helpers";
 
 import { getProductObject } from "~/library/jest/helpers/product";
-import { CustomerProductServiceUpsert } from "../../services/product";
+import { CustomerProductServiceUpsert } from "../../services/product/upsert";
 import { CustomerScheduleServiceCreate } from "../../services/schedule/create";
 import {
   CustomerProductControllerDestroy,
