+++ src/functions/customer/controllers/products/list.spec.ts
@@ -9,7 +9,7 @@ import {
 } from "~/library/jest/azure";
 
 import { getProductObject } from "~/library/jest/helpers/product";
-import { CustomerProductServiceUpsert } from "../../services/product";
+import { CustomerProductServiceUpsert } from "../../services/product/upsert";
 import { CustomerScheduleServiceCreate } from "../../services/schedule/create";
 import {
   CustomerProductsControllerList,
